Make navigateFormByDirectLink helper, rename some helper files for clarity
