+++ tests/cypress/integration/stateWorkflow/dashboard/dashboard.spec.ts
@@ -23,23 +23,23 @@ describe('dashboard', () => {
 
         // add a draft submission
         cy.startNewContractAndRatesSubmission()
-        cy.navigateForm('SAVE_DRAFT')
+        cy.navigateFormByButtonClick('SAVE_DRAFT')
 
         // a submitted submission
         cy.startNewContractAndRatesSubmission()
 
         cy.fillOutBaseContractDetails()
-        cy.navigateForm('CONTINUE')
+        cy.navigateFormByButtonClick('CONTINUE')
 
         cy.fillOutNewRateCertification()
-        cy.navigateForm('CONTINUE')
+        cy.navigateFormByButtonClick('CONTINUE')
 
         cy.fillOutStateContact()
         cy.fillOutActuaryContact()
-        cy.navigateForm('CONTINUE')
+        cy.navigateFormByButtonClick('CONTINUE')
 
         cy.fillOutSupportingDocuments()
-        cy.navigateForm('CONTINUE')
+        cy.navigateFormByButtonClick('CONTINUE')
 
         // Store submission name for reference later
         let submissionId = ''
