+++ tests/cypress/integration/stateWorkflow/stateSubmissionForm/rateDetails.spec.ts
@@ -8,26 +8,26 @@ describe('rate details', () => {
             const { pathname } = fullUrl
             const pathnameArray = pathname.split('/')
             const draftSubmissionId = pathnameArray[2]
-            cy.visit(`/submissions/${draftSubmissionId}/edit/rate-details`)
+            cy.navigateFormByDirectLink(`/submissions/${draftSubmissionId}/edit/rate-details`)
 
             // Navigate to contract details page by clicking back
-           cy.navigateForm('BACK')
+           cy.navigateFormByButtonClick('BACK')
             cy.findByRole('heading', { level: 2, name: /Contract details/ })
 
             // Navigate to rate details page
-            cy.visit(`/submissions/${draftSubmissionId}/edit/rate-details`)
+            cy.navigateFormByDirectLink(`/submissions/${draftSubmissionId}/edit/rate-details`)
 
             // Navigate to dashboard page by clicking save as draft
-            cy.navigateForm('SAVE_DRAFT')
+            cy.navigateFormByButtonClick('SAVE_DRAFT')
             cy.findByRole('heading', { level: 1, name: /Dashboard/ })
 
             // Navigate to rate details page
-            cy.visit(`/submissions/${draftSubmissionId}/edit/rate-details`)
+            cy.navigateFormByDirectLink(`/submissions/${draftSubmissionId}/edit/rate-details`)
 
             cy.fillOutNewRateCertification()
 
             // Navigate to contacts page by clicking continue
-            cy.navigateForm('CONTINUE')
+            cy.navigateFormByButtonClick('CONTINUE')
             cy.findByRole('heading', { level: 2, name: /Contacts/ })
         })
     })
@@ -41,16 +41,16 @@ describe('rate details', () => {
             const { pathname } = fullUrl
             const pathnameArray = pathname.split('/')
             const draftSubmissionId = pathnameArray[2]
-            cy.visit(`/submissions/${draftSubmissionId}/edit/rate-details`)
+            cy.navigateFormByDirectLink(`/submissions/${draftSubmissionId}/edit/rate-details`)
 
             cy.fillOutAmendmentToPriorRateCertification()
 
             // Navigate to contacts page by clicking continue
-            cy.navigateForm('CONTINUE')
+            cy.navigateFormByButtonClick('CONTINUE')
             cy.findByRole('heading', { level: 2, name: /Contacts/ })
 
             // check accessibility of filled out rate details page
-            cy.navigateForm('BACK')
+            cy.navigateFormByButtonClick('BACK')
             // Commented out to get react-scripts/webpack 5 upgrade through
             // cy.pa11y({
             //     actions: ['wait for element #form-guidance to be visible'],
@@ -69,18 +69,18 @@ describe('rate details', () => {
             const { pathname } = fullUrl
             const pathnameArray = pathname.split('/')
             const draftSubmissionId = pathnameArray[2]
-            cy.visit(`/submissions/${draftSubmissionId}/edit/rate-details`)
+            cy.navigateFormByDirectLink(`/submissions/${draftSubmissionId}/edit/rate-details`)
 
             cy.fillOutAmendmentToPriorRateCertification()
 
             // Navigate to contacts page by clicking continue
-            cy.navigateForm('CONTINUE')
+            cy.navigateFormByButtonClick('CONTINUE')
             cy.findByRole('heading', { level: 2, name: /Contacts/ })
 
             
             cy.fillOutStateContact()
             cy.fillOutActuaryContact()
-            cy.navigateForm('CONTINUE')
+            cy.navigateFormByButtonClick('CONTINUE')
 
             cy.findByRole('heading', { level: 2, name: /Supporting documents/ })
