+++ tests/cypress/support/navigateCommands.ts
@@ -1,0 +1,61 @@
+import { aliasQuery, aliasMutation } from '../utils/graphql-test-utils'
+
+type FormButtonKey =
+    | 'CONTINUE_FROM_START_NEW'
+    | 'CONTINUE'
+    | 'SAVE_DRAFT'
+    | 'BACK'
+type FormButtons = { [key in FormButtonKey]: string }
+const buttonsWithLabels: FormButtons = {
+    CONTINUE: 'Continue',
+    CONTINUE_FROM_START_NEW: 'Continue',
+    SAVE_DRAFT: 'Save as draft',
+    BACK: 'Back',
+}
+
+
+Cypress.Commands.add(
+    'navigateFormByButtonClick',
+    (buttonKey: FormButtonKey, waitForLoad = true) => {
+        cy.intercept('POST', '*/graphql', (req) => {
+            aliasQuery(req, 'indexHealthPlanPackages')
+            aliasQuery(req, 'fetchHealthPlanPackage')
+            aliasMutation(req, 'createHealthPlanPackage')
+            aliasMutation(req, 'updateHealthPlanFormData')
+        })
+        cy.findByRole('button', {
+            name: buttonsWithLabels[buttonKey],
+        }).should('not.have.attr', 'aria-disabled')
+        cy.findByRole('button', {
+            name: buttonsWithLabels[buttonKey],
+        }).safeClick()
+
+        if (buttonKey === 'SAVE_DRAFT') {
+            cy.findByTestId('dashboard-page').should('exist')
+        } else if (buttonKey === 'CONTINUE_FROM_START_NEW') {
+            if (waitForLoad) {
+                cy.wait('@createHealthPlanPackageMutation', { timeout: 50000 })
+                cy.wait('@fetchHealthPlanPackageQuery')
+            }
+            cy.findByTestId('state-submission-form-page').should('exist')
+        } else if (buttonKey === 'CONTINUE') {
+            if (waitForLoad) {
+                cy.wait('@updateHealthPlanFormDataMutation')
+            }
+            cy.findByTestId('state-submission-form-page').should('exist')
+        } else {
+            cy.findByTestId('state-submission-form-page').should('exist')
+        }
+    }
+)
+
+
+Cypress.Commands.add(
+    'navigateFormByDirectLink',
+    (url: string, waitForLoad = true) => {
+        cy.intercept('POST', '*/graphql', (req) => {
+            aliasQuery(req, 'fetchHealthPlanPackage')
+        })
+    cy.visit(url)
+    if (waitForLoad)  cy.wait('@fetchHealthPlanPackageQuery')
+    })
